Remove `asdf` & back to `n`
`asdf` is just not ready for primetime.  It's more a great idea than a
well-executed idea.  I hope it succeeds and I can switch back to it at a
later date.  But for now moving back to good ol' `n`.

Note: I'm intentionally not installing `rbenv`.  I barely do Ruby these
days and installing a version manager for it seems way overkill.  Just
going to use the one that comes pre-installed on macOS right now.