Add `Makefile` & move back to `rbenv`
Completely rewrote my Makefile from the ground up in an attempt to
actually learn Make and not cargo-cult my way through it.  I sort of
know Make now.  Not an expert but not a novice either.  So goal kind of
achieved? 🤷‍♂️

Dropped `asdf` in a prior commit
(f1a83074a85602f02032671c691b505f591aa25f).  I attempted to use the
system's ruby, but ultimately it was annoying because it required `sudo`
for all commands.

Same thing with `n` being installed in Homebrew's `bin`.  I set a custom
install location for `n` to `~/.n` so that no longer requires sudo.