Rework dotfiles based on setting up new machine (#5)
Basically an entire rethink of the dotfiles.

* I realized I no longer really needed `asdf` to manage various versions
of various languages. I just installed any languages I need via
homebrew. If I end up in a situation where I need to jump between
different versions of a language, I can deal with that then.
* Re-set-up LazyVim from scratch.
* Switch from rcup to stow because rcup seems abandoned and stow is
simpler.
* Radically simplified the `Makefile`. It had gotten away from me and
was overly clever. I set up a new machine about once every 3 years or
so. So the overkill of making it super clever was just not worth it.
* Added some sensible bullet points and references to the README..